feat(showcase/google-adk/reasoning-default): rename from reasoning-default-render and port frontend from langgraph-python
